Tighten header, remove keyboard shortcuts floating button
