+++ app/scenes/Document/components/Footer.tsx
@@ -1,6 +1,5 @@
 import styled from "styled-components";
 import type Document from "~/models/Document";
-import KeyboardShortcutsButton from "./KeyboardShortcutsButton";
 import ConnectionStatus from "./ConnectionStatus";
 import { SizeWarning } from "./SizeWarning";
 
@@ -12,14 +11,13 @@ export const Footer = ({ document }: Props) => (
   <FooterWrapper>
     <ConnectionStatus />
     <SizeWarning document={document} />
-    <KeyboardShortcutsButton />
   </FooterWrapper>
 );
 
 const FooterWrapper = styled.div`
   position: fixed;
   bottom: 12px;
-  right: 20px;
+  right: 12px;
   text-align: right;
   display: flex;
   justify-content: flex-end;
